2.6.3.0
Added option to retrieve a DataAdapter from a SQLConnector.
Added Update option for DataAdapters.

Added RestrictabkeTableEditor(DockPanel (WPF), used to display or edit Tables within a Program.
It is possible to restrict the access to definable Tables.
FnLog 2.0.2.0
Added option for integrating it into a existing Sqlite Database, which has to be connected via 
the  de.fearvel.net.SQL.Connector.SqliteConnector